requirements.txt
alembic==1.12.0
altgraph==0.17.3
aniso8601==9.0.1
attrs==23.1.0
authy==2.2.6
blinker==1.6.2
boto3==1.28.40
botocore==1.31.40
bottle==0.12.25
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
filelock==3.12.3
Flask==2.3.3
flask-buzz==2.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
flask-paginate==2022.1.8
flask-praetorian==1.4.0
flask-restx==1.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jira==3.5.2
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.0
natsort==8.4.0
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
pandas==2.1.0
passlib==1.7.4
pendulum==2.1.2
pluggy==1.3.0
protobuf==4.21.12
proxy-tools==0.1.0
py-buzz==2.1.3
pycparser==2.21
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PyJWT==2.8.0
pyotp==2.9.0
pytest==7.4.1
python-dateutil==2.8.2
pytz==2023.3
pytzdata==2020.1
pywebview==4.3.2
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rpds-py==0.10.2
s3transfer==0.6.2
schedule==1.2.0
six==1.16.0
SQLAlchemy==1.4.18
tldextract==3.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
WTForms==2.3.3
